Divide the index.js functions into logic and rendering modules
